Extract version to separate package to prevent cyclic imports
This PR is to reduce the number of changes in https://github.com/elastic/beats/pull/3681 .
