+++ libbeat/version/helper.go
@@ -1,0 +1,7 @@
+package version
+
+// GetDefaultVersion returns the current libbeat version.
+// This method is in a separate file as the version.go file is auto generated
+func GetDefaultVersion() string {
+	return defaultBeatVersion
+}
