+++ libbeat/version/version.go
@@ -1,3 +1,3 @@
-package beat
+package version
 
 const defaultBeatVersion = "6.0.0-alpha1"
